python_modules/requirements.txt
PyYAML
docopts